// src/app/real-time/models/notificacion.model.ts
export class Notificacion {

    id: string; 
    uniqueId: string; 
    content: string; 
    url: string; 
    status: string; 
    source: string; 
    sourceType: string; 
    sourceId: string; 
    sourceAlias: string; 
    targetId: string; 
    target: string; 
    targetAlias: string; 
    targetType: string; 
    targetEmail: string; 
    targetPhone: string; 
    createdAt: string; 
    createdBy: string; 
    updatedAt: string; 
    updatedBy: string; 
    
}